Re-org code, WIP of fixing the camera layers
