+++ package.json
@@ -41,6 +41,7 @@
     "@types/react-native": "^0.62.11",
     "@typescript-eslint/eslint-plugin": "^3.0.1",
     "@typescript-eslint/parser": "^3.0.1",
+    "babel-plugin-module-resolver": "^4.0.0",
     "babel-preset-expo": "~8.1.0",
     "eslint": "^7.1.0",
     "eslint-config-prettier": "^6.11.0",
