+++ src/components/App/styles.ts
@@ -1,0 +1,51 @@
+import { StyleSheet } from "react-native";
+
+export default StyleSheet.create({
+  app: {
+    // backgroundColor: "#fff",
+    backgroundColor: "transparent",
+    flex: 1
+  },
+  capture: {
+    // alignSelf: "center",
+    // backgroundColor: "#fff",
+    // borderRadius: 5,
+    // flex: 0,
+    // paddingHorizontal: 20
+  },
+  camera: {
+    // bottom: 0,
+    // left: 0,
+    // position: "absolute",
+    // right: 0,
+    // top: 0,
+    // zIndex: 0
+  },
+  cameraContainer: {
+    // backgroundColor: "red",
+    // display: "flex",
+    // height: "65%",
+    // position: "absolute",
+    // top: 0,
+    // width: "100%",
+    // zIndex: 2
+  },
+  controls: {
+    padding: 24
+  },
+  header: {
+    // color: "black",
+    // fontSize: 42,
+    // fontWeight: "900"
+  },
+  preview: {
+    // alignItems: "center",
+    // flex: 1,
+    // justifyContent: "flex-end"
+  },
+  predictions: {
+    // height: "100%",
+    // width: "100%",
+    // zIndex: 19
+  }
+});
