+++ src/components/Camera/style.ts
@@ -1,0 +1,13 @@
+import { StyleSheet } from "react-native";
+
+export const style = StyleSheet.create({
+  camera: {
+    flex: 1
+    // bottom: 0,
+    // left: 0,
+    // position: "absolute",
+    // right: 0,
+    // top: 0,
+    // zIndex: 0
+  }
+});
