+++ src/components/Controls/index.tsx
@@ -1,0 +1,7 @@
+import * as React from "react";
+import { Text, View } from "react-native";
+import styles from "./styles";
+
+export default function Controls() {
+  return <View style={styles.container}></View>;
+}
