+++ src/components/Controls/styles.ts
@@ -1,0 +1,8 @@
+import { StyleSheet } from "react-native";
+import { UNIT } from "lib/styles";
+
+export default StyleSheet.create({
+  container: {
+    padding: UNIT * 2
+  }
+});
