+++ src/components/LoadGraph/styles.ts
