+++ src/components/Readout/index.tsx
@@ -1,0 +1,7 @@
+import * as React from "react";
+import { View } from "react-native";
+import styles from "./styles";
+
+export default function Readout() {
+  return <View style={styles.readout} />;
+}
