+++ src/components/Readout/styles.ts
@@ -1,0 +1,8 @@
+import { StyleSheet } from "react-native";
+
+export default StyleSheet.create({
+  readout: {
+    backgroundColor: "transparent",
+    flex: 1
+  }
+});
