+++ src/lib/styles.ts
@@ -1,0 +1,1 @@
+export const UNIT = 16;
